requirements.txt
fastapi
uvicorn
openai
faiss-cpu
sentence-transformers
PyMuPDF
numpy
